Refactor JSON serialization methods in JsonUtils
- Replaced all instances of JsonUtils.Serialize with JsonUtils.ToJson for consistency across the codebase.
- Updated unit tests to reflect the change in serialization method.
- Removed unused interfaces and classes related to request and response models to streamline the code.
- Bumped version number to 0.2.2 in the project file.
