--- ReflectorNet/src/Model/Request/IRequestID.cs
@@ -1,8 +1,0 @@
-
-namespace com.IvanMurzak.ReflectorNet.Model
-{
-    public interface IRequestID
-    {
-        string RequestID { get; set; }
-    }
-}
