--- ReflectorNet/src/Model/Request/Resource/Content/IRequestResourceContent.cs
@@ -1,9 +1,0 @@
-using System;
-
-namespace com.IvanMurzak.ReflectorNet.Model
-{
-    public interface IRequestResourceContent : IRequestID, IDisposable
-    {
-        public string Uri { get; set; }
-    }
-}
