--- ReflectorNet/src/Model/Request/Resource/List/IRequestListResources.cs
@@ -1,9 +1,0 @@
-using System;
-
-namespace com.IvanMurzak.ReflectorNet.Model
-{
-    public interface IRequestListResources : IRequestID, IDisposable
-    {
-        public string? Cursor { get; set; }
-    }
-}
