--- ReflectorNet/src/Model/Request/Resource/Template/IRequestListResourceTemplates.cs
@@ -1,8 +1,0 @@
-using System;
-
-namespace com.IvanMurzak.ReflectorNet.Model
-{
-    public interface IRequestListResourceTemplates : IRequestID, IDisposable
-    {
-    }
-}
