--- ReflectorNet/src/Model/Request/Tool/List/IRequestListTool.cs
@@ -1,8 +1,0 @@
-using System;
-
-namespace com.IvanMurzak.ReflectorNet.Model
-{
-    public interface IRequestListTool : IRequestID, IDisposable
-    {
-    }
-}
