--- ReflectorNet/src/Model/Request/Tool/List/RequestListTool.cs
@@ -1,23 +1,0 @@
-using System;
-
-namespace com.IvanMurzak.ReflectorNet.Model
-{
-    public class RequestListTool : IRequestListTool
-    {
-        public string RequestID { get; set; } = Guid.NewGuid().ToString();
-
-        // Empty constructor for JSON deserialization
-        public RequestListTool() { }
-
-        // Overloaded constructor to set RequestID
-        public RequestListTool(string requestId)
-        {
-            RequestID = requestId ?? throw new ArgumentNullException(nameof(requestId));
-        }
-
-        public virtual void Dispose()
-        {
-        }
-        ~RequestListTool() => Dispose();
-    }
-}
