--- ReflectorNet/src/Model/Response/Resource/Content/ResponseResourceContentExtensions.cs
@@ -1,24 +1,0 @@
-using System;
-using Microsoft.Extensions.Logging;
-
-namespace com.IvanMurzak.ReflectorNet.Model
-{
-    public static class ResponseResourceContentExtensions
-    {
-        public static ResponseResourceContent[] Log(this ResponseResourceContent[] target, ILogger logger, Exception? ex = null)
-        {
-            if (!logger.IsEnabled(LogLevel.Information))
-                return target;
-
-            foreach (var item in target)
-                logger.LogInformation(ex, "Resource: {0}", item.uri);
-
-            return target;
-        }
-
-
-        public static IResponseData<ResponseResourceContent[]> Pack(this ResponseResourceContent[] target, string requestId, string? message = null)
-            => ResponseData<ResponseResourceContent[]>.Success(requestId, message ?? "List Tool execution completed.")
-                .SetData(target);
-    }
-}
