--- ReflectorNet/src/Model/Response/Tool/Call/IResponseCallTool.cs
@@ -1,10 +1,0 @@
-using System.Collections.Generic;
-
-namespace com.IvanMurzak.ReflectorNet.Model
-{
-    public interface IResponseCallTool
-    {
-        bool IsError { get; set; }
-        List<ResponseCallToolContent> Content { get; set; }
-    }
-}
