+++ ReflectorNet/src/Reflector/Reflector.CallMethod.cs
@@ -104,7 +104,7 @@ namespace com.IvanMurzak.ReflectorNet
                     : methodWrapper.Invoke();
 
                 var result = task.Result;
-                return $"[Success] Execution result:\n```json\n{JsonUtils.Serialize(result)}\n```";
+                return $"[Success] Execution result:\n```json\n{JsonUtils.ToJson(result)}\n```";
             };
 
             if (executeInMainThread)
